fix: melhorando a visibidade dos viewModels
